requirements.txt
pytest
playwright==1.58.0
pytest-playwright==0.5.0
python-dotenv
greenlet==3.1.1